+++ view/base/web/js/source/magento-plugin.js
@@ -1,21 +1,15 @@
-import ConfigManager from './config_manager'
-
-// Unlike our other plugins, the Magento plugin doesn't use the addressfinder-webpage-tools npm package. This is because Magento doesn't have
-// support for npm. Instead, we take the file that the npm package outputs, and copy the code into the addressfinder-webpage-tools javascript file. 
-import { PageManager, MutationManager } from './addressfinder-webpage-tools'
+import {PageManager, MutationManager} from '@addressfinder/addressfinder-webpage-tools'
 
 export default class MagentoPlugin {
-  constructor(widgetConfig) {
+  constructor(widgetConfig, formsConfig) {
     this.widgetConfig = widgetConfig
-    this.widgetOptions = this._parseWidgetOptions(widgetConfig.options)
+    this.formsConfig = formsConfig || []
+    this.widgetOptions = widgetConfig.options || {}
 
-    this.version = "1.2.1"
+    this.version = "1.4.0"
 
-    // Manages the mapping of the form configurations to the DOM. 
+    // Manages the mapping of the form configurations to the DOM.
     this.PageManager = null
-
-    // Manages the form configurations, and creates any dynamic forms
-    this.ConfigManager = null
 
     this._initPlugin()
 
@@ -25,22 +19,10 @@ export default class MagentoPlugin {
 
   mutationEventHandler() {
     // When the form mutates, reload our form configurations, and reload the form helpers in the page manager.
-    let addressFormConfigurations = this.ConfigManager.load()
     if (this.PageManager) {
-      this.PageManager.reload(addressFormConfigurations)
+      this.PageManager.reload(this.formsConfig)
     }
   }
-
-  _parseWidgetOptions(options) {
-    try {
-      return JSON.parse(options);
-    } catch (error) {
-      if (this.debugMode) {
-        console.warn("Widget options ignored. They must be in valid JSON format");
-      }
-      return {};
-    }
-  };
 
   _initPlugin() {
 
@@ -49,10 +31,9 @@ export default class MagentoPlugin {
       auKey: this.widgetConfig.key,
       nzWidgetOptions: this.widgetOptions,
       auWidgetOptions: this.widgetOptions,
-      debug: this.widgetConfig.debug || false
+      debug: this.widgetConfig.debug || false,
+      defaultCountry: this.widgetConfig.default_search_country
     }
-
-    this.ConfigManager = new ConfigManager()
 
     // Watches for any mutations to the DOM, so we can reload our configurations when something changes.
     new MutationManager({
@@ -62,7 +43,7 @@ export default class MagentoPlugin {
     })
 
     this.PageManager = new PageManager({
-      addressFormConfigurations: this.ConfigManager.load(),
+      addressFormConfigurations: this.formsConfig,
       widgetConfig: widgetConfig,
       // When an address is selected dispatch this event on all the updated form fields. This tells the store the fields have been changed.
       formFieldChangeEventToDispatch: 'change',
@@ -70,7 +51,15 @@ export default class MagentoPlugin {
       countryChangeEventToListenFor: 'change'
     })
 
+    this._setVersionNumbers()
+
     window.AddressFinder._magentoPlugin = this.PageManager
+  }
+
+  _setVersionNumbers() {
+    // rename webpage tools version from 'version' to 'webpageToolsVersion'
+    this.PageManager['webpageToolsVersion'] = this.PageManager.version
+    this.PageManager.version = this.version
   }
 
   /*