+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# AddressFinder 1.4.0 (January 6th 2020)
+* Fetch state mappings and payment methods server side.
+* Offically remove support for Magento 2.0
+
+# AddressFinder 1.3.0 (December 9th 2019)
+* Utilise NPM
+* Upgrades JavaScript dependencies
+* Refactor project structure
+
+# AddressFinder 1.2.5 (October 25th 2019)
+* Added default static country to default to if the country select field is not present.
+
+# AddressFinder 1.2.4 (October 16th 2019)
+* Readme.md updates
+
+# AddressFinder 1.2.3 (October 15th 2019)
+* Differentiate between Magento version and webpage tools version
+
+# AddressFinder 1.2.2 (October 14th 2019)
+* Update to webpage tools v1.6.0
+
 # AddressFinder 1.2.1 (June 13th 2019)
 * Fix fault with incorrectly named widgetOptions